Fixed NoneType bug with pars like CHOP, SOP, PYTHON, etc
The way I was checking to see if a par existed, was to getattr() the par, however this could succeed and still return None if the par was a type that referenced an operator, and returned None if no valid operator was plugged in.

Revised the code slightly to use a different method for validation.